move dsnot
